Procedure for DataScience Data Collation
This is my procedure which I've created to clean some of the data and merge together. I've input the data using SSIS (the powerpoint has a screenshot of SSIS.

Please see attached Questions variables, which I'm using as a tracking sheet as well. 

My next step is to create a view to impute some of the 'Unknowns', where I can. For instance I plan to impute age based on the mean age of Countries (I believe 2017 has missing data for age and therefore needs to be imputed).

I plan to create the views late this week and start building the visualisations as well.